added ScringCodableMap wrapper type and computed values for privatized wrapped types to allow for type coercion / type change on decoded values

+++ Sources/RetroAPIPackage/Protocols/AchievementProtocol.swift
@@ -28,7 +28,7 @@ public protocol Achievement {
 
 public protocol ExtendedAchievement {
     
-    var trueRatio:Int { get }
+    var trueRatio:Double { get }
     var author:String { get }
     var dateModified:Date { get }
     var dateCreated:Date { get }
